Add links to Readme.md
the pdf uses a absolute link. 
This is problematic for users who clone this repo but only this way it is possible to get a direct clickable link to the pdf. 
This way the user can download or view the pdf in a local pdf viewer and not via the github built-in pdf viewer, which I wanted.